修复糖尿病体检流程中吸烟页面崩溃,体检报告身高,体重单位重复问题

+++ app/src/main/java/com/example/han/referralproject/health/model/DetectResult.java
@@ -86,10 +86,10 @@ public class DetectResult implements Serializable {
          * lowPressure : 0
          */
 
-        public int highPressure;
-        public int lowPressure;
+        public float highPressure;
+        public float lowPressure;
 
-        public RightHypertensionBean(int highPressure, int lowPressure) {
+        public RightHypertensionBean(float highPressure, float lowPressure) {
             this.highPressure = highPressure;
             this.lowPressure = lowPressure;
         }
